feat(postgre): Uses postgre instead of mongoDB
